Calculate National Emission Inventory (NEI) covariates

NEI data comprises multiple csv files where emissions of 50+ pollutants are recorded at county level. With raw data files, this function will join a combined table of NEI data and county boundary, then perform a spatial join to the sites.

calc_nei(
  path = "./input/nei/",
  sites,
  id_col = "site_id",
  year = 2017,
  county_shp = NULL,
  sites_epsg = "EPSG:4326"
)

Arguments

path

character(1). Path to the directory with NEI CSV files

sites

stdt/sf/SpatVector/data.frame. Unique sites. See convert_stobj_to_stdt for details of stdt

id_col

character(1). Unique site identifier column name. Default is "site_id". It is no more than a placeholder in this function

year

integer(1). Data year. Currently only accepts c(2017, 2020)

county_shp

character(1). Path to county boundary file.

sites_epsg

character(1). Coordinate system of sites.

Value

A data.frame object.

Author

Insang Song, Ranadeep Daw