download_data: a wrapper function for download covariate sources

download_data: a wrapper function for download covariate sources

download_data(
  dataset_name = c("aqs", "ecoregion", "geos", "gmted", "koppen", "koppengeiger",
    "merra2", "merra", "narr_monolevel", "modis", "narr_p_levels", "nlcd", "noaa",
    "sedac_groads", "sedac_population", "groads", "population", "plevels", "p_levels",
    "monolevel", "hms", "smoke"),
  directory_to_save = NULL,
  data_download_acknowledgement = FALSE,
  ...
)

Arguments

dataset_name

character(1). Dataset to download.

directory_to_save

character(1). Directory to save / unzip (if zip files are downloaded) data.

data_download_acknowledgement

logical(1). By setting `= TRUE` the user acknowledge that the data downloaded using this function may be very large and use lots of machine storage and memory.

...

Arguments passed to each download function.

Note

- All download function names are in download_*_data formats

See also

Author

Insang Song