Download global elevation data from the Global Multi-resolution Terrain Elevation Data (GMTED2010).

The download_gmted_data() function acesses and downloads Global Multi-resolution Terrain Elevation Data (GMTED2010) from U.S. Geological Survey and National Geospatial-Intelligence Agency.

download_gmted_data(
  statistic = c("Breakline Emphasis", "Systematic Subsample", "Median Statistic",
    "Minimum Statistic", "Mean Statistic", "Maximum Statistic",
    "Standard Deviation Statistic"),
  resolution = c("7.5 arc-seconds", "15 arc-seconds", "30 arc-seconds"),
  directory_to_download = "./input/gmted/",
  directory_to_save = "./input/gmted/",
  data_download_acknowledgement = FALSE,
  unzip = TRUE,
  remove_zip = FALSE,
  download = FALSE,
  remove_command = FALSE
)

Arguments

statistic

character(1). Available statistics include "Breakline Emphasis", "Systematic Subsample", "Median Statistic", "Minimum Statistic", "Mean Statistic", "Maximum Statistic", and "Standard Deviation Statistic".

resolution

character(1). Available resolutions include "7.5 arc-seconds", "15 arc-seconds", and "30 arc-seconds".

directory_to_download

character(1). Directory to download zip files from Global Multi-resolution Terrain Elevation Data (GMTED2010).

directory_to_save

character(1). Directory to decompress zip files.

data_download_acknowledgement

logical(1). By setting TRUE the user acknowledge that the data downloaded using this function may be very large and use lots of machine storage and memory.

unzip

logical(1). Unzip zip files. Default is TRUE.

remove_zip

logical(1). Remove zip file from directory_to_download. Default is FALSE.

download

logical(1). FALSE will generate a *.txt file containing all download commands. By setting TRUE the function will download all of the requested data files.

remove_command

logical(1). Remove (TRUE) or keep (FALSE) the text file containing download commands. Default is FALSE.

Value

NULL;

Author

Mitchell Manware, Insang Song