download_sedac_population_data: download UN WPP-Adjusted population density data from NASA Socioeconomic Data and Applications Center (SEDAC).

The `download_sedac_population_data()` function accesses and downloads population density data from the National Aeronatuics and Space Administration's (NASA) [UN WPP-Adjusted Population Density, v4.11] (https://sedac.ciesin.columbia.edu/data/set/gpw-v4-population-density-adjuste d-to-2015-unwpp-country-totals-rev11).

download_sedac_population_data(
  year = "2020",
  data_format = "GeoTIFF",
  data_resolution = "60 minute",
  directory_to_download = "./input/sedac_population/",
  directory_to_save = "./input/sedac_population/",
  data_download_acknowledgement = FALSE,
  download = FALSE,
  unzip = TRUE,
  remove_zip = FALSE,
  remove_command = FALSE
)

Arguments

year

character(1). Available years are 2000, 2005, 2010, 2015, and 2020, or "all" for all years.

data_format

character(1). Individual year data can be downloaded as "ASCII" or "GeoTIFF". "all" years is downloaded as "netCDF".

data_resolution

character(1). Available resolutions are 30 second (~1km), 2.5 minute (~5km), 15 minute (~30km), 30 minute (~55km), and 60 minute (~110km).

directory_to_download

character(1). Directory to download zip files from NASA UN WPP-Adjusted Population Density, v4.11.

directory_to_save

character(1). Directory to decompress zip files.

data_download_acknowledgement

logical(1). By setting `= TRUE` the user acknowledge that the data downloaded using this function may be very large and use lots of machine storage and memory.

download

logical(1). `= FALSE` will generate a `.txt` file containing all download commands. By setting `= TRUE` the function will download all of the requested data files.

unzip

logical(1). Unzip zip files. Default = `TRUE`.

remove_zip

logical(1). Remove zip files from directory_to_download. Default = `FALSE`.

remove_command

logical(1). Remove (TRUE) or keep (FALSE) the text file containing download commands.

Value

NULL;

Author

Mitchell Manware, Insang Song