Fit base learner

This function fits a selected base learner, which is supposed to be one of random forest, XGBoost, and convolutional neural network, using one of spatiotemporal cross-validation approaches.

Usage

fit_base_learner(
  covars,
  learner = c("randomforest", "xgboost", "cnn"),
  dependent_name = "pm2.5",
  independent_name,
  cv_mode = c("lolo", "loto", "lolto", "random", "lblo", "lbto", "lblto"),
  cv_fold = 5L,
  return_full_object = FALSE
)

Arguments

covars

stdt. See convert_stobj_to_stdt

learner

character(1). Currently one of 'randomforest', 'xgboost', and 'cnn'

dependent_name

character(1). Name of the dependent variable.

independent_name

character(1). Names of independent variables.

cv_mode

character(1). One of 'lolo (leave-one-location-out)', 'loto (leave-one-time-out)', 'lolto (leave-one-location-time-out)', 'lblo (leave-block-location-out)', 'lbto (leave-block-time-out)', 'lblto (leave-block-location-time-out)', and 'random (full random selection)'

cv_fold

integer(1). Number of folds for cross-validation.

return_full_object

logical(1). TRUE will return the object whose class is dependent on the package in use, whereas FALSE will only return the prediction with spatial/temporal information. Full object is potentially helpful for debugging.

Value

data.frame

Author

Insang Song