Get mosaicked or merged raster from multiple MODIS hdf files

Get mosaicked or merged raster from multiple MODIS hdf files

modis_get_vrt(
  paths,
  product = c("MOD11A1", "MOD13A2", "MOD09GA", "MCD19A2"),
  date_in = NULL,
  regex_sds = NULL,
  foo = "mean"
)

Arguments

paths

character. Full list of hdf file paths. preferably a recursive search result from list.files.

product

character(1). Product code of MODIS. Should be one of c("MOD11A1", "MOD13A2", "MOD09GA", "MCD19A2")

date_in

character(1). date to query. Should be in "YYYY-MM-DD" format.

regex_sds

character(1). subdataset names to extract. Should conform to regular expression. See regex for details. Default is NULL, which means that the subdataset names are automatically selected based on product value.

foo

Function name or custom function to aggregate overlapping cell values. See fun description in tapp for details.

Value

A SpatRaster object.

Author

Insang Song