A single-date MODIS worker for parallelization

modis_worker operates at six MODIS/VIIRS products (MOD11A1, MOD13A2, MOD06_L2, VNP46A2, MOD09GA, and MCD19A2) on a daily basis. Given that the raw hdf files are downloaded from NASA, standard file names include a data retrieval date flag starting with A. Leveraging that piece of information, the function will select files of scope on the date of interest. Please note that this function does not provide a function to filter swaths or tiles, so it is strongly recommended to check and pre-filter the file names at users' discretion.

modis_worker(
  raster,
  date,
  sites_in = NULL,
  name_extracted = NULL,
  product = c("MOD11A1", "MOD13A2", "MOD06_L2", "VNP46A2", "MOD09GA", "MCD19A2"),
  fun_summary_raster = "mean",
  id_col = "site_id",
  radius = 0L
)

Arguments

raster

SpatRaster.

date

Date(1). date to query.

sites_in

SpatVector/sf/sftime object. AQS sites.

name_extracted

character. Names of calculated covariates.

product

character(1). Product code of MODIS. Should be one of c('MOD11A1', 'MOD13A2', 'MOD06_L2', 'VNP46A2', 'MOD09GA', 'MCD19A2')

fun_summary_raster

function. Summary function for multilayer rasters. Passed to foo. See also exact_extract

id_col

character(1). Field name where unique site identifiers are stored. Default is "site_id"

radius

numeric. Radius to buffer.

Value

A data.frame object.

Author

Insang Song